Move taggable interface from tag bundle into component
