+++ src/Badger/Bundle/UserBundle/Entity/User.php
@@ -3,7 +3,7 @@
 namespace Badger\UserBundle\Entity;
 
 use Badger\Component\Tag\Model\TagInterface;
-use Badger\TagBundle\Taggable\TaggableInterface;
+use Badger\Component\Tag\Taggable\TaggableInterface;
 use Doctrine\Common\Collections\ArrayCollection;
 use FOS\UserBundle\Model\User as BaseUser;
