+++ src/Badger/Component/Game/Model/BadgeInterface.php
@@ -2,7 +2,7 @@
 
 namespace Badger\Component\Game\Model;
 
-use Badger\TagBundle\Taggable\TaggableInterface;
+use Badger\Component\Tag\Taggable\TaggableInterface;
 use Symfony\Component\HttpFoundation\File\UploadedFile;
 
 /**
